+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 - Allow converting a single frame of a multiframe image (#579)
 - Add a convert endpoint to the Girder plugin (#578)
+- Added support for creating Aperio svs files (#580)
 
 ### Improvements
 - More untiled tiff files are handles by the bioformats reader (#569)